requirements.txt
PySide6
requests
python-dotenv
pytest
pytest-qt
pyinstaller
